travis: upgrade ESP8266/ESP32 Arduino cores
and fix compile errors

+++ src/app/unix-time.cpp
@@ -117,7 +117,10 @@ UnixTimeClass::processData(const ndn::DataLite& data, uint64_t endpointId)
   uint64_t timeOffset = timestamp - m_lastRequest * 1000;
   uint64_t now = timeOffset + millis() * 1000;
   if (m_timeOffset == TIMEOFFSET_INTEG) {
-    struct timeval tv = { .tv_sec = now / 1000000, .tv_usec = now % 1000000 };
+    struct timeval tv = {
+      .tv_sec = static_cast<unsigned long>(now / 1000000),
+      .tv_usec = static_cast<unsigned long>(now % 1000000),
+    };
     settimeofday(&tv, nullptr);
   }
   else {
